cloud/cdk/baseline_cdk/resources/cfnres/iot_cacert/requirements.txt
# boto3 # provided by lambda container
crhelper==2.0.6
pyopenssl==19.1.0